% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{extract_gff_attribute}
\alias{extract_gff_attribute}
\title{Extract a named attribute from GFF attribute strings}
\usage{
extract_gff_attribute(attributes, key, is_gtf)
}
\description{
Extract a named attribute from GFF attribute strings
}
\keyword{internal}
