% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{get_mcl_bin}
\alias{get_mcl_bin}
\title{Get path to an MCL tool binary}
\usage{
get_mcl_bin(tool)
}
\arguments{
\item{tool}{One of "mcl", "mcxload", "mcxdump"}
}
\description{
Get path to an MCL tool binary
}
\keyword{internal}
