% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_output.R
\name{read_anchorpoints}
\alias{read_anchorpoints}
\title{Read the anchorpoints table}
\usage{
read_anchorpoints(output_dir)
}
\arguments{
\item{output_dir}{Path to the i-ADHoRe output directory}
}
\value{
A data frame
}
\description{
Reads `anchorpoints.txt` from an i-ADHoRe output directory.
}
\details{
Columns: id, multiplicon, basecluster, gene_x, gene_y, coord_x, coord_y,
is_real_anchorpoint
}
