% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_output.R
\name{read_baseclusters}
\alias{read_baseclusters}
\title{Read the baseclusters table}
\usage{
read_baseclusters(output_dir)
}
\arguments{
\item{output_dir}{Path to the i-ADHoRe output directory}
}
\value{
A data frame
}
\description{
Reads `baseclusters.txt` from an i-ADHoRe output directory.
}
\details{
Columns: id, multiplicon, number_of_anchorpoints, orientation,
was_twisted, random_probability
}
