% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_output.R
\name{read_cloud_anchorpoints}
\alias{read_cloud_anchorpoints}
\title{Read the cloud anchorpoints table}
\usage{
read_cloud_anchorpoints(output_dir)
}
\arguments{
\item{output_dir}{Path to the i-ADHoRe output directory}
}
\value{
A data frame
}
\description{
Reads `cloudAP.txt` from an i-ADHoRe output directory (cloud/hybrid mode
only).
}
\details{
Columns: CloudID, gene_x, gene_y, coord_x, coord_y
}
