% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_input.R
\name{read_fasta_ids}
\alias{read_fasta_ids}
\title{Read gene IDs from a FASTA file (part of header before first space)}
\usage{
read_fasta_ids(fasta_file)
}
\description{
Read gene IDs from a FASTA file (part of header before first space)
}
\keyword{internal}
