% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_output.R
\name{read_genes}
\alias{read_genes}
\title{Read the genes table}
\usage{
read_genes(output_dir)
}
\arguments{
\item{output_dir}{Path to the i-ADHoRe output directory}
}
\value{
A data frame
}
\description{
Reads `genes.txt` from an i-ADHoRe output directory.
}
\details{
Columns: id, genome, list, coordinate, orientation, remapped_coordinate,
is_tandem, is_tandem_representative, tandem_representative, remapped
}
