% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_output.R
\name{read_iadhore_table}
\alias{read_iadhore_table}
\title{Read a tab-delimited i-ADHoRe output table}
\usage{
read_iadhore_table(file)
}
\description{
Read a tab-delimited i-ADHoRe output table
}
\keyword{internal}
