% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_output.R
\name{read_list_elements}
\alias{read_list_elements}
\title{Read the list_elements table}
\usage{
read_list_elements(output_dir)
}
\arguments{
\item{output_dir}{Path to the i-ADHoRe output directory}
}
\value{
A data frame
}
\description{
Reads `list_elements.txt` from an i-ADHoRe output directory.
}
\details{
Columns: id, segment, gene, position, orientation
}
