% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_output.R
\name{read_multiplicons}
\alias{read_multiplicons}
\title{Read the multiplicons table}
\usage{
read_multiplicons(output_dir)
}
\arguments{
\item{output_dir}{Path to the i-ADHoRe output directory}
}
\value{
A data frame
}
\description{
Reads `multiplicons.txt` from an i-ADHoRe output directory.
}
\details{
Columns: id, genome_x, list_x, parent, genome_y, list_y, level,
number_of_anchorpoints, profile_length, begin_x, end_x, begin_y, end_y,
is_redundant
}
