% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_output.R
\name{read_segments}
\alias{read_segments}
\title{Read the segments table}
\usage{
read_segments(output_dir)
}
\arguments{
\item{output_dir}{Path to the i-ADHoRe output directory}
}
\value{
A data frame
}
\description{
Reads `segments.txt` from an i-ADHoRe output directory.
}
\details{
Columns: id, multiplicon, genome, list, first, last, order
}
